feat(createApp): added styles and modal and now works properly creating an Application in the database
